imx: hab: Fix compilation warning
Fix the following:
In function ‘imx_hab_authenticate_image’:
warning: format ‘%lx’ expects argument of type ‘long unsigned int’,
but argument 3 has type ‘uint32_t’ {aka ‘unsigned int’}

+++ arch/arm/mach-imx/hab.c
@@ -737,7 +737,7 @@ int imx_hab_authenticate_image(uint32_t ddr_start, uint32_t image_size,
 	status = hab_rvt_check_target(HAB_TGT_MEMORY, (void *)(ulong)ddr_start, bytes);
 	if (status != HAB_SUCCESS) {
 		printf("HAB check target 0x%08x-0x%08lx fail\n",
-		       ddr_start, ddr_start + bytes);
+		       ddr_start, (long unsigned int) ddr_start + bytes);
 		goto hab_exit_failure_print_status;
 	}
 #ifdef DEBUG
